22 bloque buscador

See merge request ctd/hispanos/proyecto-integrador-1/proyecto-integrador-0523/1021pt-c3/equipo-10!79

+++ ui/src/components/ListCategory/index.jsx
@@ -9,6 +9,7 @@ import { ModalCategory } from '../Modal/ModalCategory';
 import { AiFillDelete, AiFillEdit } from 'react-icons/ai';
 //import { AiFillDelete } from 'react-icons/ai';
 import { GrAdd } from 'react-icons/gr';
+import Swal from "sweetalert2";
 import './ListCategory.css'
 
 export const ListCategory = () => {
@@ -54,9 +55,29 @@ export const ListCategory = () => {
 
   const handleDeleteCategorie = async (categoryId) => {
     try {
-      await deleteCategory(categoryId);
+      const result = await Swal.fire({
+        title: 'Estas seguro que desea eliminar una categoría.',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#6D9886',
+        cancelButtonColor: '#ED2B2A',
+        confirmButtonText: 'Ok'
+      });
+
+      if (result.isConfirmed) {
+        await deleteCategory(categoryId);
+        Swal.fire({
+          title: 'Tu categoría ha sido eliminada.',
+          icon: 'success',
+          confirmButtonColor: '#6D9886',
+          confirmButtonText: 'Ok'
+        });
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.close();
+      }
     } catch (error) {
-      console.error("Error deleting category:", error);
+      console.error("Error al borrar categoría:", error);
+      Swal.fire("Error", "Se ha producido un error al eliminar la categoría.", "error");
     }
   };
 
@@ -72,15 +93,32 @@ export const ListCategory = () => {
           type: actions.UPDATE_CATEGORY,
           payload: updatedCategory,
         });
+        Swal.fire({
+          title: "Categoría actualizada",
+          confirmButtonColor: '#6D9886',
+          icon: "success"
+        }).then(() => {
+          window.location.reload();
+        });
       } else {
         await addCategory(updatedCategory);
+        Swal.fire({
+          title: "Categoría añadida",
+          confirmButtonColor: '#6D9886',
+          icon: "success"
+        }).then(() => {
+          window.location.reload();
+        });
       }
       closeModal();
       reset();
-
-      window.location.reload();
     } catch (error) {
       console.error("Error adding/updating category:", error);
+      Swal.fire(
+        "Error",
+        "Se ha producido un error al añadir/actualizar la categoría.",
+        "error"
+      );
     }
   };
